# Conflicts:
#	src/main/java/com/sparta/finalproject/domain/attendance/repository/AttendanceRepository.java
#	src/main/java/com/sparta/finalproject/domain/child/controller/ChildController.java
#	src/main/java/com/sparta/finalproject/domain/child/service/ChildService.java
#	src/main/resources/import.sql

+++ src/main/java/com/sparta/finalproject/domain/gallery/service/ImagePostService.java
@@ -9,10 +9,13 @@ import com.sparta.finalproject.domain.gallery.entity.Image;
 import com.sparta.finalproject.domain.gallery.entity.ImagePost;
 import com.sparta.finalproject.domain.gallery.repository.ImagePostRepository;
 import com.sparta.finalproject.domain.gallery.repository.ImageRepository;
+import com.sparta.finalproject.domain.user.entity.User;
 import com.sparta.finalproject.global.dto.GlobalResponseDto;
+import com.sparta.finalproject.global.enumType.UserRoleEnum;
 import com.sparta.finalproject.global.response.CustomStatusCode;
 import com.sparta.finalproject.global.response.exceptionType.ClassroomException;
 import com.sparta.finalproject.global.response.exceptionType.ImagePostException;
+import com.sparta.finalproject.global.response.exceptionType.UserException;
 import com.sparta.finalproject.infra.s3.S3Service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -39,7 +42,10 @@ public class ImagePostService {
     private static final int PAGE_NUMBER = 15;
 
     @Transactional
-    public GlobalResponseDto imagePostAdd(Long classroom_id, ImagePostRequestDto imagePostRequestDto) throws IOException {
+    public GlobalResponseDto addImagePost(Long classroom_id, ImagePostRequestDto imagePostRequestDto, User user) throws IOException {
+        if(user.getRole() != UserRoleEnum.ADMIN){
+            throw new UserException(CustomStatusCode.UNAUTHORIZED_USER);
+        }
         Classroom classroom = classroomRepository.findById(classroom_id).orElseThrow(
                 () -> new ClassroomException(CustomStatusCode.CLASSROOM_NOT_FOUND)
         );
@@ -55,7 +61,7 @@ public class ImagePostService {
     }
 
     @Transactional(readOnly = true)
-    public GlobalResponseDto imagePostFind(Long imagePostId){
+    public GlobalResponseDto findImagePost(Long imagePostId){
         ImagePost imagePost = imagePostRepository.findById(imagePostId).orElseThrow(
                 () -> new ImagePostException(CustomStatusCode.IMAGE_POST_NOT_FOUND)
         );
@@ -68,9 +74,12 @@ public class ImagePostService {
     }
 
     @Transactional
-    public GlobalResponseDto imagePostDelete(Long imagePostId) {
-        List<Image> imagePostList = imageRepository.findAllByImagePostId(imagePostId);
-        for (Image image : imagePostList) {
+    public GlobalResponseDto deleteImagePost(Long imagePostId, User user) {
+        if(user.getRole() != UserRoleEnum.ADMIN){
+            throw new UserException(CustomStatusCode.UNAUTHORIZED_USER);
+        }
+        List<Image> imageList = imageRepository.findAllByImagePostId(imagePostId);
+        for (Image image : imageList) {
             s3Service.deleteFile(image.getImageUrl().substring(63));
         }
         try {
@@ -82,7 +91,7 @@ public class ImagePostService {
         return GlobalResponseDto.from(CustomStatusCode.DELETE_IMAGE_POST_SUCCESS);
     }
 
-    public GlobalResponseDto imagePostPageFind(Long classroomId, String start, String end, String keyword, int page, boolean isAsc) {
+    public GlobalResponseDto findImagePostPage(Long classroomId, String start, String end, String keyword, int page, boolean isAsc) {
         Sort.Direction direction = isAsc ? Sort.Direction.ASC : Sort.Direction.DESC;
         Sort sort = Sort.by(direction, "id");
         Pageable pageable = PageRequest.of(page, PAGE_NUMBER, sort);
@@ -98,5 +107,30 @@ public class ImagePostService {
         return GlobalResponseDto.of(CustomStatusCode.FIND_IMAGE_POST_PAGE_SUCCESS,
                 GalleryPageResponseDto.of(imagePostCount, responseDtoList));
 
+    }
+
+    @Transactional
+    public GlobalResponseDto modifyImagePost(Long imagePostId, ImagePostRequestDto imagePostRequestDto, User user) throws IOException{
+        if(user.getRole() != UserRoleEnum.ADMIN){
+            throw new UserException(CustomStatusCode.UNAUTHORIZED_USER);
+        }
+        ImagePost imagePost = imagePostRepository.findById(imagePostId).orElseThrow(
+                () -> new ImagePostException(CustomStatusCode.IMAGE_POST_NOT_FOUND)
+        );
+        List<Image> beforeImageList = imageRepository.findAllByImagePostId(imagePostId);
+        for (Image image : beforeImageList) {
+            s3Service.deleteFile(image.getImageUrl().substring(63));
+        }
+        imagePost.update(imagePostRequestDto);
+        imageRepository.deleteAllByImagePostId(imagePostId);
+        List<MultipartFile> afterImageList = imagePostRequestDto.getImageList();
+        if(afterImageList != null){
+            s3Service.upload(afterImageList, "gallery", imagePost);
+        }
+        Image image = imageRepository.findFirstByImagePost(imagePost);
+        List<String> imageUrlList = new ArrayList<>();
+        imageUrlList.add(image.getImageUrl());
+        return GlobalResponseDto.of(CustomStatusCode.MODIFY_IMAGE_POST_SUCCESS,
+                ImagePostResponseDto.of(imagePost, imageUrlList));
     }
 }
